Remove custom run script and add grain function
I used to do a lot of work in the bin command that was kinda fiddly and
didn't really work. Now I am deciding to go to a more natural approach
to how salt stack wants you to work. This commit adds a _grains folder
to add custom functions to add grains. Grains where the only thing that
the /bin/run.sh script really was doing anymore so without that the
script just covers up salt stack and I would rather expose all of salt
when building dotfiles