Add options to set user permissions for actions
When running salt stack on arch it has to run as sudo. This meant that
after running the commands I didn't have permissions to folders. I would
constantly run around and `chown` things to set it as me. This change
should make that issue go away since salt can set the permissions up as
its doing the actions.