[PASS] Add --no-throw-keyids to gpg options
My default gpg config doesn't put the recipient in the encoded message.
This causes issues over on andriod where I am trying to read the
secrets. OpenPGP on android needs the recipient in the payload. Since
pass will always be me encoding messages to myself its fine to pass in
the override to gpg to ensure that other password managers can derive
the key.