[NVIM] Fix Errors in lsp module
I once again made the mistake of letting old files stay behind and cover
up mistakes. The new files weren't actually getting written. This adds a
new hack that I found in the packer.nvim repo that runs the packer sync
command and closes automatically on exit. This gives me the ability to
delete all the old config files of each run of titan without any of the
painpoints of having to sync plugins afterwards