[AWE] Refactor player widget
Moved to its own file and try to hook in new bling playerctl logic which
I am having all sorts of issues getting to work. Going to keep this
scaffolding in place but not have the widget hook into signals. This way
at least titan isn't broken on other machines when I apply the next
highstate.