[SALT] Make user (orlando) configurable through static grain
This removes most of the places I was hard coding orlando and instead
reads from a static grain set in the minion config. This will be more
helpful for others who want to copy and paste my config and have things
work. Or even for myself if I decide to put a more professional name on
my system.... That probably won't happen though