[TMUX] Fix launching tmuxintator templates
The change to having rally uppercase the session names had some
side effects. When trying to look up a tmuxinator template it was
looking for the name by uppercase. This change adds an additional
variable that is the session name vs the template name. Also updated old
templates to uppercase the session name manually.