Reorganize zsh startup files
I have recently learned that the zshenv is started before the zprofile
file. This wasn't an issue in linux but in wsl where the login shell was
also the shell I would work in the env was wrong. Reorganizing what is
in the files and making zshenv the entrypoint file with most important
env.