[NVIM] Configure telescope live grep to add to qflist
Enter normally goes to the item under the cursor which always frustrated
me because I would lose the other results. Now enter will add entries to
the quickfix and take me to the first one.

Going to flip the mappings so that the old grep way is at _ instead of /