[NVIM] Properly set global variables in Lazy
I was noticing that global variables were not being applied in plugins
after switching to lazy. I have found out that the issue is that the
plugin was running before the `config` callback was ran. This meant that
the variables were getting set too late.

By moving to `init` the variables always get set and before plugin is
processed.