[ZSH] Remove .zprofile and move all zsh to .config
The load order of my files will now be as follows:

* `/etc/zsh/zshenv`
  * Sets the ZDOTDIR to `~/.config/zsh` to keep all files into the home
    folder
* `~/.config/zsh/.zshenv`
  * Sets all environment variables
* `~/.config/zsh/.zshrc`
  * All interactive commands

Changes here involved creating `/etc/zsh/zshenv` so that `ZDOTDIR` can
be set early in the chain. Disolving the `.zprofile` into the other
files. And lastly placing the GPG tty in the correct spot (turns out it
needed interactive shells).